+++ taskforce/apps/users/src/app/auth/auth.http
@@ -1,35 +1,10 @@
-# Create a new user
-POST http://localhost:3333/api/auth/register HTTP/1.1
-Content-Type: application/json
-
-{
-  "userName": "ПетрПетрович",
-  "email": "academy@test.ru",
-  "city": "Москва",
-  "password": "123456",
-  "dateBirth": "1993-10-18",
-  "role": "Customer"
-}
-###
-
 # Login user by email and password
 POST http://localhost:3333/api/auth/login HTTP/1.1
 Content-Type: application/json
 
 {
   "email": "test123@test.ru",
-  "password": "asdadadad123123"
-}
-###
-
-# Update user's password
-PUT http://localhost:3333/api/auth/password HTTP/1.1
-Content-Type: application/json
-
-{
-  "email": "test123@test.ru",
-  "currentPassword": "asdadadad123123",
-  "newPassword": "123456789"
+  "password": "123456"
 }
 ###
 
@@ -55,6 +30,7 @@ Content-Type: application/json
 ###
 
 # Get user by ID
-GET http://localhost:3333/api/auth/73316262-20ea-4b22-8134-c5ed263c0219 HTTP/1.1
+GET http://localhost:3333/api/user/73316262-20ea-4b22-8134-c5ed263c0219 HTTP/1.1
+Authorization: Bearer {'eyJzdWIiOiI2Mzk3MTNhMmIzYjk2ZDA5OTg0OWQ5MjMiLCJlbWFpbCI6InVzZXJAbm90Zm91bmQubG9jYWwiLCJyb2xlIjoiY2xpZW50IiwibmFtZSI6Iktla3MiLCJpYXQiOjE2NzExOTUyMDAsImV4cCI6MTY3MTE5NTI2MH0.jJz24iZhn7YetFhGwA2Dw3uJq0huxwq5OFtBy-80I3w'}
 
 ###
