+++ taskforce/apps/users/src/app/auth/dto/login-user.dto.ts
@@ -1,31 +1,11 @@
-import { ApiProperty } from '@nestjs/swagger';
-import {
-  AuthUserError,
-  MAX_LENGTH_PASSWORD,
-  MIN_LENGTH_PASSWORD,
-  UserApiDescription,
-} from '../auth.constant';
-import { IsEmail, Length } from 'class-validator';
+import { PickType } from '@nestjs/swagger';
+import CreateUserDto from '../../user/dto/create-user.dto';
 
-export class LoginUserDto {
-  @ApiProperty({
-    description: UserApiDescription.Email,
-    example: 'admin@admin.ru',
-  })
-  @IsEmail(
-    {},
-    {
-      message: AuthUserError.EmailNotValid,
-    }
-  )
-  public email: string;
-
-  @ApiProperty({
-    description: UserApiDescription.Password,
-    example: '123456',
-  })
-  @Length(MIN_LENGTH_PASSWORD, MAX_LENGTH_PASSWORD, {
-    message: AuthUserError.PasswordIsWrong,
-  })
-  public password: string;
+export class LoginUserDto extends PickType(CreateUserDto, [
+  'email',
+  'password',
+] as const) {
+  constructor() {
+    super();
+  }
 }
