+++ taskforce/apps/users/src/app/auth/rdo/logged-user.rdo.ts
@@ -1,6 +1,6 @@
 import { Expose, Transform } from 'class-transformer';
 import { ApiProperty } from '@nestjs/swagger';
-import { UserApiDescription } from '../auth.constant';
+import { UserApiDescription } from '../../user/user.constant';
 
 export class LoggedUserRdo {
   @ApiProperty({
