+++ taskforce/apps/users/src/app/user/dto/update-user-password.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { UserApiDescription } from '../auth.constant';
+import { UserApiDescription } from '../user.constant';
 
 // TODO: валидация?
