+++ taskforce/apps/users/src/app/user/user.http
@@ -1,0 +1,24 @@
+# Create a new user
+POST http://localhost:3333/api/user/register HTTP/1.1
+Content-Type: application/json
+
+{
+  "name": "Keks",
+  "email": "test@test.ru",
+  "city": "Москва",
+  "password": "123456",
+  "dateBirth": "2003-12-22",
+  "role": "customer"
+}
+###
+
+# Update user's password
+PATCH http://localhost:3333/api/user/password HTTP/1.1
+Content-Type: application/json
+
+{
+  "email": "test@test.ru",
+  "currentPassword": "123456",
+  "newPassword": "234567"
+}
+###
