+++ taskforce/apps/users/src/app/user/user.module.ts
@@ -2,12 +2,23 @@ import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { UserModel, UserSchema } from './user.model';
 import UserRepository from './user.repository';
+import { JwtService } from '@nestjs/jwt';
+import { UserService } from './user.service';
+import { AuthService } from '../auth/auth.service';
+import { UserController } from './user.controller';
 
 @Module({
   imports: [
-    MongooseModule.forFeature([{ name: UserModel.name, schema: UserSchema }]),
+    // AuthModule,
+    MongooseModule.forFeature([
+      {
+        name: UserModel.name,
+        schema: UserSchema,
+      },
+    ]),
   ],
-  providers: [UserRepository],
+  providers: [UserRepository, UserService, AuthService, JwtService],
   exports: [UserRepository],
+  controllers: [UserController],
 })
 export class UserModule {}
