+++ taskforce/apps/users/src/app/validators/age-validator.ts
@@ -3,7 +3,7 @@ import {
   ValidatorConstraintInterface,
 } from 'class-validator';
 import * as dayjs from 'dayjs';
-import { MIN_USER_AGE } from '../auth/auth.constant';
+import { MIN_USER_AGE } from '../user/user.constant';
 
 @ValidatorConstraint({ name: 'Invalid Age', async: false })
 export class AgeValidator implements ValidatorConstraintInterface {
