+++ taskforce/libs/core/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/helpers';
 export * from './lib/crud-repository.interface';
+export * from './lib/jwt-auth.guard';
