+++ taskforce/package.json
@@ -43,6 +43,7 @@
     "@nrwl/node": "15.2.4",
     "@nrwl/workspace": "15.2.4",
     "@types/jest": "28.1.1",
+    "@types/multer": "^1.4.7",
     "@types/node": "18.7.1",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
